requirements.txt
requests~=2.32.4
python-dotenv~=1.1.1
gradio~=5.41.1
python-dateutil~=2.9.0.post0
pandas~=2.3.1
aiohttp~=3.12.15
matplotlib~=3.10.5
numpy~=2.2.6
pyarrow~=21.0.0
ipython~=8.37.0
torch~=2.7.1+cu128
faiss-cpu
sentence-transformers
streamlit
scikit-learn